Refactor hiring module
- move application and opening entities to related modules
- move staking_policy methods to staking_policy module
