+++ src/hiring/mod.rs
@@ -1,12 +1,13 @@
+mod application;
 mod opening;
 mod staking_policy;
-mod application;
 
-pub use opening::*;
 pub use application::*;
+pub use opening::*;
 pub use staking_policy::*;
 
-
-
-
-
+#[derive(Eq, PartialEq, Clone, Debug)]
+pub enum StakePurpose {
+    Role,
+    Application,
+}
